Add shared Google auth helper
