+++ InsightMate/InsightMateApp/Resources/py/gmail_reader.py
@@ -1,31 +1,8 @@
 from __future__ import print_function
-import os.path
 from email.header import decode_header, make_header
-from google.oauth2.credentials import Credentials
-from google_auth_oauthlib.flow import InstalledAppFlow
-from google.auth.transport.requests import Request
 from googleapiclient.discovery import build
 
-SCOPES = ['https://www.googleapis.com/auth/gmail.readonly',
-          'https://www.googleapis.com/auth/calendar.readonly']
-
-TOKEN_FILE = 'token.json'
-CREDENTIALS_FILE = 'credentials.json'
-
-
-def get_credentials():
-    creds = None
-    if os.path.exists(TOKEN_FILE):
-        creds = Credentials.from_authorized_user_file(TOKEN_FILE, SCOPES)
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(CREDENTIALS_FILE, SCOPES)
-            creds = flow.run_local_server(port=0)
-        with open(TOKEN_FILE, 'w') as token:
-            token.write(creds.to_json())
-    return creds
+from google_auth import get_credentials
 
 
 def fetch_unread_email():
